+++ Android_Development_With_Flutter/Flutter_project/cloud_photos/lib/AuthPages/login.dart
@@ -1,10 +1,9 @@
-import 'package:cloud_photos/Screen/HomeScreen.dart';
+import 'package:cloud_photos/Screen/home_screen.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_signin_button/button_list.dart';
 import 'package:flutter_signin_button/button_view.dart';
 import '../Authentication/auth.dart';
-// import 'home.dart';
 import 'otp.dart';
 
 class LoginPage extends StatefulWidget {
@@ -24,12 +23,14 @@ class _LoginPageState extends State<LoginPage> {
       name = user.displayName.toString();
       email = user.email.toString();
       imgurl = user.photoURL.toString();
-      Navigator.pushReplacement(
-        context,
-        MaterialPageRoute(
-          builder: (context) => const HomeScreen(),
-        ),
-      );
+      WidgetsBinding.instance!.addPostFrameCallback((_) {
+        Navigator.pushReplacement(
+          context,
+          MaterialPageRoute(
+            builder: (context) => const HomeScreen(),
+          ),
+        );
+      });
     }
   }
 
@@ -52,74 +53,75 @@ class _LoginPageState extends State<LoginPage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        appBar: AppBar(
-          title: const Text("Login"),
-        ),
-        body: SafeArea(
-          child: SingleChildScrollView(
-            child: Container(
-              padding: const EdgeInsets.all(32),
-              child: Form(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.center,
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    TextFormField(
-                      controller: _phonecontroller,
-                      decoration: const InputDecoration(
-                        label: Text("Enter Mobile Number"),
-                        labelStyle: TextStyle(
-                          fontSize: 20.0,
-                        ),
-                        enabledBorder: OutlineInputBorder(
-                          borderSide: BorderSide(width: 0.0),
-                        ),
-                        border: OutlineInputBorder(),
-                        errorStyle:
-                            TextStyle(color: Colors.redAccent, fontSize: 15),
+      appBar: AppBar(
+        title: const Text("Login"),
+      ),
+      body: SafeArea(
+        child: SingleChildScrollView(
+          child: Container(
+            padding: const EdgeInsets.all(32),
+            child: Form(
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.center,
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  TextFormField(
+                    controller: _phonecontroller,
+                    decoration: const InputDecoration(
+                      label: Text("Enter Mobile Number"),
+                      labelStyle: TextStyle(
+                        fontSize: 20.0,
                       ),
-                      validator: (value) {
-                        if (value!.length < 10) {
-                          return 'Please valid Phone Number';
-                        }
-                        return null;
-                      },
+                      enabledBorder: OutlineInputBorder(
+                        borderSide: BorderSide(width: 0.0),
+                      ),
+                      border: OutlineInputBorder(),
+                      errorStyle:
+                          TextStyle(color: Colors.redAccent, fontSize: 15),
                     ),
-                    const SizedBox(height: 18),
-                    ElevatedButton(
-                      child: const Text("Send OTP"),
-                      onPressed: () {
-                        Navigator.of(context).push(
-                          MaterialPageRoute(
-                            builder: (context) => OTPScreen(
-                              phone: _phonecontroller.text,
-                            ),
+                    validator: (value) {
+                      if (value!.length < 10) {
+                        return 'Please valid Phone Number';
+                      }
+                      return null;
+                    },
+                  ),
+                  const SizedBox(height: 18),
+                  ElevatedButton(
+                    child: const Text("Send OTP"),
+                    onPressed: () {
+                      Navigator.of(context).push(
+                        MaterialPageRoute(
+                          builder: (context) => OTPScreen(
+                            phone: _phonecontroller.text,
                           ),
-                        );
-                      },
+                        ),
+                      );
+                    },
+                  ),
+                  const Text(
+                    "OR",
+                    style: TextStyle(fontSize: 23),
+                  ),
+                  Center(
+                    child: Column(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: [
+                        SignInButton(
+                          Buttons.Google,
+                          onPressed: () {
+                            signInMethod();
+                          },
+                        )
+                      ],
                     ),
-                    const Text(
-                      "OR",
-                      style: TextStyle(fontSize: 23),
-                    ),
-                    Center(
-                      child: Column(
-                        mainAxisAlignment: MainAxisAlignment.center,
-                        children: [
-                          SignInButton(
-                            Buttons.Google,
-                            onPressed: () {
-                              signInMethod();
-                            },
-                          )
-                        ],
-                      ),
-                    ),
-                  ],
-                ),
+                  ),
+                ],
               ),
             ),
           ),
-        ));
+        ),
+      ),
+    );
   }
 }
