+++ Android_Development_With_Flutter/Flutter_project/cloud_photos/lib/main.dart
@@ -1,8 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'AuthPages/login.dart';
-// import 'Firebase/home.dart';
-// import 'Screen/HomeScreen.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
