+++ Android_Development_With_Flutter/Flutter_project/cloud_photos/pubspec.yaml
@@ -27,16 +27,19 @@ environment:
 # the latest version available on pub.dev. To see which dependencies have newer
 # versions available, run `flutter pub outdated`.
 dependencies:
+  cloud_firestore: ^3.1.0
   cupertino_icons: ^1.0.2
   firebase_auth: ^3.1.1
   firebase_core: ^1.6.0
+  firebase_storage: ^10.1.0
   flutter:
     sdk: flutter
+  flutter_advanced_drawer: ^1.2.4
   flutter_signin_button: ^2.0.0
-  google_sign_in: ^5.1.0
-  pinput: ^1.2.2
   google_fonts: ^2.1.0
+  google_sign_in: ^5.1.0
   image_picker: ^0.8.4+4
+  pinput: ^1.2.2
 
 dev_dependencies:
   flutter_lints: ^1.0.0
